Adding reference state calculation for finite elasticity
Temporary approach to implementing reference state calculation which swaps geometric and dependent field pointers in finite elasticity routines.
While this method works, the correct way to implement this is to specify which field will be used for the dependent and geometric field explicitly in the users program script.
This temp approach is only committed to compare solutions for testing.

Note that in EQUATIONS_INTERPOLATION_INITIALISE, FIELD_GEOMETRIC_GENERAL_TYPE needs to be used (not implement in this commit see Trac item 3342).
IF(EQUATIONS%INTERPOLATION%DEPENDENT_FIELD%TYPE==FIELD_GEOMETRIC_TYPE.OR. &
  & EQUATIONS%INTERPOLATION%DEPENDENT_FIELD%TYPE==FIELD_FIBRE_TYPE) THEN
    CALL FIELD_INTERPOLATED_POINTS_METRICS_INITIALISE(EQUATIONS%INTERPOLATION%DEPENDENT_INTERP_POINT, &
      & EQUATIONS%INTERPOLATION%DEPENDENT_INTERP_POINT_METRICS,ERR,ERROR,*999)
ENDIF

and the check for the number of coordinate dimensions in FIELD_INTERPOLATED_POINTS_METRICS_INITIALISE needs to be updated to handle the FIELD_GEOMETRIC_GENERAL_TYPE
IF(NUMBER_OF_X_DIMENSIONS==SIZE(INTERPOLATED_POINT%VALUES,1)) THEN

Tracker Item 3510